requirements.txt
openai-whisper
transformers
torch
librosa
pandas
nltk
scikit-learn
vaderSentiment
bitsandbytes
accelerate
streamlit
boto3
requests
python-dotenv
watchdog
gTTS
tqdm
